requirements.txt
Flask
Pillow
tweepy
gunicorn
